Singletons in C++ should not be created on the Heap
The correct way to create a singleton in c++, since c++11 is to
use a static variable, and because this variable will not be
destructed when the function goes out of scope. The usage of
'new' is not just wrong, but it's also slower, and adds false
positives to tools that looks for memory leaks.

Also, pass the flag CppOwnership so the Qml will not try to
delete this object