Don't use [] to look for an item in the map.
If the item doesn't exists this will add an item to the map, so
it should only be used when we are certain that an item is there.
Also, adding an item to the map will not only create a new QPair,
but will also allocate memory in the heap and stack.

Use find to look for items.